refactor(orchestrator): extract CLI logic from main.rs
